+++ routes/movies.js
@@ -1,5 +1,5 @@
 const router = require('express').Router();
-const { createMovieValidation, deleteMovieValidation } = require('../middlewares/validation');
+const { createMovieValidation, deleteMovieValidation } = require('../validation/validation');
 
 const {
   getMovies, createMovie, deleteMovie,
