+++ routes/users.js
@@ -2,7 +2,7 @@ const router = require('express').Router();
 const {
   updateUser, getCurrentUser,
 } = require('../controllers/users');
-const { updatedUserValidation } = require('../middlewares/validation');
+const { updatedUserValidation } = require('../validation/validation');
 
 router.get('/users/me', getCurrentUser);
 router.patch('/users/me', updatedUserValidation, updateUser);
