--- utils/allowedCors.js
@@ -1,6 +1,0 @@
-const allowedCors = [
-  'http://localhost:3001',
-  'http://localhost:3000',
-];
-
-module.exports = allowedCors;
