+++ utils/constants.js
@@ -13,4 +13,11 @@ const error = {
   incorrectUserIdError: 'Передан некорректный id пользователя',
 };
 
-module.exports = error;
+const urlRegEx = /^(http|https):\/\/(www\.)?[-a-zA-Z0-9@:%_.~#?&=]{2,256}\.[a-z]{2,4}\b(\/[-a-zA-Z0-9@:%_.~#?&=]*)?/i;
+
+const allowedCors = [
+  'http://localhost:3001',
+  'http://localhost:3000',
+];
+
+module.exports = { error, urlRegEx, allowedCors };
