--- utils/urlValidation.js
@@ -1,3 +1,0 @@
-const urlRegEx = /^(http|https):\/\/(www\.)?[-a-zA-Z0-9@:%_.~#?&=]{2,256}\.[a-z]{2,4}\b(\/[-a-zA-Z0-9@:%_.~#?&=]*)?/i;
-
-module.exports = urlRegEx;
